Adding anchorClassName option and moving anchor
I've found that sometimes (maybe always?) it's preferable to have the
anchor before the text rather than wrapping it. This way you can offset
it with css or do whatever.. there's no real need to use it to style the
text itself, you should already have classes/styles on the text.

Also added an anchorClassName option in case you don't like toc-anchor
or maybe it conflicts. Options are always nice!